// app/api/types.ts
export interface ApiResponse<T = void> {
  success: boolean;
  data?: T;
  error?: string;
}
// ...
export type SleepLogResponse = Omit<
  SleepLog,
  "startTime" | "endTime" | "createdAt" | "updatedAt" | "deletedAt"
> & {
  startTime: string;
  endTime: string | null;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
};

export interface SleepLogCreate {
  babyId: string;
  startTime: string;
  endTime?: string;
  duration?: number;
  type: SleepType;
  location?: string;
  quality?: SleepQuality;
}